Fixed Mesannine twister to use UInt32

+++ src/MersenneTwister.h
@@ -59,10 +59,12 @@
 #include <ctime>
 #include <cmath>
 
+#include "Globals.h"
+
 class MTRand {
 // Data
 public:
-	typedef long uint32;  // unsigned integer type, at least 32 bits
+	typedef UInt32 uint32;  // unsigned integer type, at least 32 bits
 	
 	enum { N = 624 };       // length of state vector
 	enum { SAVE = N + 1 };  // length of array for save()
